+++ aiccs-20191015/src/main/java/com/aliyun/aiccs20191015/models/GetAgentIndexRealTimeRequest.java
@@ -14,10 +14,10 @@ public class GetAgentIndexRealTimeRequest extends TeaModel {
     public Integer currentPage;
 
     @NameInMap("DepIds")
-    public java.util.List<Integer> depIds;
+    public java.util.List<Long> depIds;
 
     @NameInMap("GroupIds")
-    public java.util.List<Integer> groupIds;
+    public java.util.List<Long> groupIds;
 
     public static GetAgentIndexRealTimeRequest build(java.util.Map<String, ?> map) throws Exception {
         GetAgentIndexRealTimeRequest self = new GetAgentIndexRealTimeRequest();
@@ -48,19 +48,19 @@ public class GetAgentIndexRealTimeRequest extends TeaModel {
         return this.currentPage;
     }
 
-    public GetAgentIndexRealTimeRequest setDepIds(java.util.List<Integer> depIds) {
+    public GetAgentIndexRealTimeRequest setDepIds(java.util.List<Long> depIds) {
         this.depIds = depIds;
         return this;
     }
-    public java.util.List<Integer> getDepIds() {
+    public java.util.List<Long> getDepIds() {
         return this.depIds;
     }
 
-    public GetAgentIndexRealTimeRequest setGroupIds(java.util.List<Integer> groupIds) {
+    public GetAgentIndexRealTimeRequest setGroupIds(java.util.List<Long> groupIds) {
         this.groupIds = groupIds;
         return this;
     }
-    public java.util.List<Integer> getGroupIds() {
+    public java.util.List<Long> getGroupIds() {
         return this.groupIds;
     }
